[FE] feat: 마이페이지 접근시 회원 정보 조회 요청 추가 (#316)
* feat: 마이페이지 접근시 회원 정보 조회 요청 추가

* chore: 디자인 시스템 버전 업데이트
