+++ frontend/src/hooks/auth/index.ts
@@ -1,1 +1,2 @@
 export { default as useAuth } from './useAuth';
+export { default as useMember } from './useMember';
